requirements.txt
numpy==2.2.5
opencv_python==4.11.0.86
pysetup
pyserial